character deletion added
