+++ src/app/library/_components/DeleteDialog.tsx
@@ -9,34 +9,35 @@ import {
 } from "~/components/ui/dialog";
 import { Input } from "~/components/ui/input";
 import { Label } from "~/components/ui/label";
-import { type Book } from "~/types";
 
-export default function DeleteBookDialog({
-  book,
+export default function DeleteDialog<T extends { name: string; id: number }>({
+  item,
   closeDialog,
-  deleteBook,
+  deleteItem,
 }: {
-  book: Book;
+  item: T;
   closeDialog: () => void;
-  deleteBook: (bookId: number) => void;
+  deleteItem: (itemId: number) => void;
 }) {
   const [input, setInput] = useState("");
 
   function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
-    if (input === book.name) {
-      deleteBook(book.id);
+    if (input === item.name) {
+      deleteItem(item.id);
       closeDialog();
     }
   }
 
+  console.log(item);
+
   return (
     <DialogContent className="rounded-none border-red-600">
       <DialogHeader>
-        <DialogTitle>Delete {book.name}</DialogTitle>
+        <DialogTitle>Delete {item.name}</DialogTitle>
       </DialogHeader>
       <form className="flex flex-col gap-2" onSubmit={handleSubmit}>
-        <Label>Type &quot;{book.name}&quot; to confirm</Label>
+        <Label>Type &quot;{item.name}&quot; to confirm</Label>
         <div className="flex gap-2">
           <Input
             value={input}
